twitch-bot/requirements.txt
nba_api==1.9.0
python-dotenv==1.1.0
